requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
multidict==6.0.4
numpy==1.26.2
pandas==2.1.4
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
